Add sql script to link author names and epersons
Publications only contain author first and last name, but there
is no link to extra info of the author, or disambiguation, or matching
if typo in the name, or misspelled.

This script performs a pretty blunt search using the first and last
names in publication to match to any existing eperson, that at least
contains an email.

In the future we will require users to login using ORCID as a minimum