Add sql and data to extract comments
These comments are swarmed by spam. However there are a few interesting
conversations worth to keep.
In order to filter the spam, the following condition is applied to
category something as spam:

```
A: comment_content has links (contains 'http')
B: comment_author has 0 submissions (not an author in the journal)
if (A and B) is true, the comment is filtered out.
```

Related:
https://github.com/InsightSoftwareConsortium/InsightJournal/issues/19