Add python script to generate a valid publication_comments.json
The current sql query has a where clause (to filter out the spam)
that only generates output when the publication meet the conditions.
This is not expected for the merge_json_data_by_publication_id script,
which expect that every publication is listed, with empty comments:
`comments: []`.

The python script is an ad-hoc solution to generate a valid
publication_comments.json with all the publications.

I am sure there is a SQL way to do this, but easier with python.