Add script extract_citation_list_from_pdf
Script to extract a citation list from a pdf.

It uses [CERMINE](https://github.com/CeON/CERMINE), as recommended
by crossref. It is used in production by OpenAIRE. The results are
pretty good, specially when compared with other pdf parsers (anystyle,
etc).

The output of CERMINE is a xml file `.cermxml` with the parsed
information.
We are only interested in the citation list.

This script runs CERMINE in a given input folder. The CERMINE jar file
has to be downloaded and provided to this script with the arg: `--cermine-path`
A `${pdf_name}.cermxml` and `${pdf_name}_citations.json` are generated
as output in the input_folder.

The idea is to run this script in the data from:
https://data.kitware.com/#collection/5cb75e388d777f072b41e8db

We have to decide if running it for each revision (1,2,3, subfolders), or
only the last revision.